Move api-docs-builder config to scripts
