+++ package.json
@@ -15,7 +15,7 @@
     "release:publish:dry-run": "pnpm publish --recursive --tag latest --registry=\"http://localhost:4873/\"",
     "release:tag": "node scripts/releaseTag.mjs",
     "docs:api": "rimraf --glob ./docs/pages/**/api-docs ./docs/pages/**/api && pnpm docs:api:build",
-    "docs:api:build": "tsx --tsconfig ./scripts/buidApiDocs/tsconfig.json ./scripts/buidApiDocs/index.ts",
+    "docs:api:build": "tsx --tsconfig ./scripts/buildApiDocs/tsconfig.json ./scripts/buildApiDocs/index.ts",
     "docs:build": "pnpm --filter docs build",
     "docs:build-sw": "pnpm --filter docs build-sw",
     "docs:deploy": "pnpm --filter docs run deploy",
@@ -83,7 +83,6 @@
     "@babel/preset-typescript": "^7.23.3",
     "@babel/register": "^7.23.7",
     "@mnajdova/enzyme-adapter-react-18": "^0.2.0",
-    "@mui-internal/api-docs-builder-core": "workspace:^",
     "@mui-internal/docs-utils": "^1.0.2",
     "@mui/internal-markdown": "^1.0.0",
     "@mui/internal-scripts": "^1.0.1",
@@ -117,6 +116,7 @@
     "cpy-cli": "^5.0.0",
     "cross-env": "^7.0.3",
     "danger": "^11.3.1",
+    "docs": "workspace:^",
     "enzyme": "^3.11.0",
     "eslint": "^8.57.0",
     "eslint-config-airbnb": "^19.0.4",
