--- packages/api-docs-builder-core/.eslintrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-  rules: {
-    'import/no-default-export': 'error',
-    'import/prefer-default-export': 'off',
-  },
-};
