--- packages/api-docs-builder-core/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "@mui-internal/api-docs-builder-core",
-  "version": "1.0.0",
-  "description": "MUI Core-specific settings for API docs generator",
-  "private": "true",
-  "main": "./index.ts",
-  "scripts": {
-    "test": "cd ../../ && cross-env NODE_ENV=test mocha 'packages/api-docs-builder-core/**/*.test.{js,ts,tsx}'",
-    "typescript": "tsc -p tsconfig.json"
-  },
-  "dependencies": {
-    "@mui/internal-markdown": "^1.0.0",
-    "@mui/monorepo": "github:mui/material-ui#5e4e18a10085b7e65955ca11ba6c93c838003e0c",
-    "docs": "workspace:^",
-    "lodash": "^4.17.21"
-  },
-  "devDependencies": {
-    "@types/babel__core": "^7.20.5",
-    "@types/babel__traverse": "^7.20.5",
-    "@types/chai": "^4.3.12",
-    "@types/doctrine": "^0.0.9",
-    "@types/mdast": "4.0.3",
-    "@types/mocha": "^10.0.6",
-    "@types/node": "^18.19.18",
-    "@types/sinon": "^17.0.3",
-    "ast-types": "^0.14.2",
-    "chai": "^4.4.1",
-    "sinon": "^17.0.1",
-    "typescript": "^5.3.3"
-  }
-}
