--- packages/api-docs-builder-core/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": ["./**/*.ts"],
-  "exclude": ["./node_modules"]
-}
