+++ scripts/buildApiDocs/config/baseUi/generateApiLinks.ts
