+++ scripts/buildApiDocs/config/baseUi/generateBaseUiApiPages.ts
