+++ scripts/buildApiDocs/config/baseUi/getBaseUiHookInfo.ts
