+++ scripts/buildApiDocs/config/baseUi/projectSettings.ts
