+++ scripts/buildApiDocs/tsconfig.json
